services/requirements.txt
# requirements.txt

# Web Framework & Server
Flask
gunicorn
uvicorn  # Required ASGI worker for Gunicorn/Flask async

# Google Cloud & API Libraries
google-auth               # Explicit auth helpers
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-storage
google-generativeai       # Gemini API Client

# Data Handling & HTTP Requests
pandas
requests

# Third-party APIs
sec-api

# Utilities
python-dotenv             # Loads environment variables from .env file
tenacity                  # Retry logic helper